requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
credentials==1.1
et-xmlfile==1.1.0
fuzzywuzzy==0.18.0
idna==3.4
numpy==1.24.2
openpyxl==3.1.2
pandas==1.5.3
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
soupsieve==2.4
urllib3==1.26.15
openai
python-dotenv
langchain
tiktoken
chromadb
flask
flask-sqlalchemy 
mysql-connector-python
flask-login
sqlalchemy
PyMySQL
xlrd
gunicorn
mysqlclient
boto3
python-Levenshtein
PyGithub